src/requirements.txt
psutil
matplotlib
pycryptodome
pytest
numpy
